Github hook handler differentiates between events.
The github payloads for different events are sometimes ambigious. There is now easy field to look at to tell exactly what type of event is being pushed, so we need to add a bit of logic to peek inside the payload to figure out what type of event it is before handling it. This commit identifies a 'push' event properly.